pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helmiesagents"
version = "0.2.0"
description = "HelmiesAgents - competitive AI agent runtime with workflows, memory, gateways, RBAC, benchmarks, and enterprise controls"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Wael Helmi"}
]
license = {text = "MIT"}
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "pyyaml>=6.0.2",
  "pyjwt>=2.9.0",
  "cryptography>=43.0.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
helmiesagents = "helmiesagents.cli:app"

[tool.setuptools]
packages = ["helmiesagents"]

[tool.setuptools.package-data]
helmiesagents = ["web/index.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
